create RawMailContext
